# apps/web/lib/api-types-openapi.json
{"openapi": "3.1.0", "info": {"title": "Trading Algo API", "version": "0.1.0"}, "paths": {"/api/login": {"post": {"tags": ["auth"], "summary": "Login", "operationId": "login_api_login_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/LoginIn"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TokenOut"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/state": {"get": {"tags": ["api"], "summary": "Get State", "operationId": "get_state_api_state_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/StateOut"}}}}}, "security": [{"HTTPBearer": []}]}}, "/api/pnl": {"get": {"tags": ["api"], "summary": "Get Pnl", "operationId": "get_pnl_api_pnl_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PnLOut"}}}}}, "security": [{"HTTPBearer": []}]}}, "/api/positions": {"get": {"tags": ["api"], "summary": "Get Positions", "operationId": "get_positions_api_positions_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/PositionOut"}, "type": "array", "title": "Response Get Positions Api Positions Get"}}}}}, "security": [{"HTTPBearer": []}]}}, "/api/orders": {"get": {"tags": ["api"], "summary": "Get Orders", "operationId": "get_orders_api_orders_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/OrderOut"}, "type": "array", "title": "Response Get Orders Api Orders Get"}}}}}, "security": [{"HTTPBearer": []}]}}, "/api/trades": {"get": {"tags": ["api"], "summary": "Get Trades", "operationId": "get_trades_api_trades_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/TradeOut"}, "type": "array", "title": "Response Get Trades Api Trades Get"}}}}}, "security": [{"HTTPBearer": []}]}}, "/api/chain": {"get": {"tags": ["api"], "summary": "Get Chain", "operationId": "get_chain_api_chain_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ChainOut"}}}}}, "security": [{"HTTPBearer": []}]}}, "/api/control/{command}": {"post": {"tags": ["api"], "summary": "Control", "operationId": "control_api_control__command__post", "security": [{"HTTPBearer": []}], "parameters": [{"name": "command", "in": "path", "required": true, "schema": {"type": "string", "title": "Command"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ControlOut"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/config": {"get": {"tags": ["api"], "summary": "Get Config", "operationId": "get_config_api_config_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"additionalProperties": true, "type": "object", "title": "Response Get Config Api Config Get"}}}}}, "security": [{"HTTPBearer": []}]}, "put": {"tags": ["api"], "summary": "Put Config", "operationId": "put_config_api_config_put", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ConfigIn"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"additionalProperties": true, "type": "object", "title": "Response Put Config Api Config Put"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"HTTPBearer": []}]}}, "/api/stream": {"get": {"tags": ["api"], "summary": "Stream", "operationId": "stream_api_stream_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"HTTPBearer": []}]}}, "/health": {"get": {"summary": "Health", "operationId": "health_health_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"additionalProperties": {"type": "string"}, "type": "object", "title": "Response Health Health Get"}}}}}}}}, "components": {"schemas": {"ChainOut": {"properties": {"ce_oi_total": {"type": "integer", "title": "Ce Oi Total"}, "pe_oi_total": {"type": "integer", "title": "Pe Oi Total"}, "selected_side": {"type": "string", "title": "Selected Side"}, "per_strike": {"items": {"$ref": "#/components/schemas/ChainStrikeOut"}, "type": "array", "title": "Per Strike"}}, "type": "object", "required": ["ce_oi_total", "pe_oi_total", "selected_side", "per_strike"], "title": "ChainOut"}, "ChainStrikeOut": {"properties": {"strike": {"type": "number", "title": "Strike"}, "ce_oi": {"type": "integer", "title": "Ce Oi"}, "ce_ltp": {"type": "number", "title": "Ce Ltp"}, "pe_oi": {"type": "integer", "title": "Pe Oi"}, "pe_ltp": {"type": "number", "title": "Pe Ltp"}}, "type": "object", "required": ["strike", "ce_oi", "ce_ltp", "pe_oi", "pe_ltp"], "title": "ChainStrikeOut"}, "ConfigIn": {"properties": {"updates": {"additionalProperties": true, "type": "object", "title": "Updates"}}, "type": "object", "required": ["updates"], "title": "ConfigIn"}, "ControlOut": {"properties": {"ok": {"type": "boolean", "title": "Ok"}, "command": {"type": "string", "title": "Command"}}, "type": "object", "required": ["ok", "command"], "title": "ControlOut"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "LoginIn": {"properties": {"password": {"type": "string", "title": "Password"}}, "type": "object", "required": ["password"], "title": "LoginIn"}, "OrderOut": {"properties": {"order_id": {"type": "string", "title": "Order Id"}, "symbol": {"type": "string", "title": "Symbol"}, "side": {"type": "string", "title": "Side"}, "quantity": {"type": "integer", "title": "Quantity"}, "filled_quantity": {"type": "integer", "title": "Filled Quantity"}, "price": {"type": "string", "title": "Price"}, "order_type": {"type": "string", "title": "Order Type"}, "product": {"type": "string", "title": "Product"}, "status": {"type": "string", "title": "Status"}, "order_time": {"type": "string", "title": "Order Time"}}, "type": "object", "required": ["order_id", "symbol", "side", "quantity", "filled_quantity", "price", "order_type", "product", "status", "order_time"], "title": "OrderOut"}, "PnLOut": {"properties": {"total_realized": {"type": "number", "title": "Total Realized"}, "total_buy_value": {"type": "number", "title": "Total Buy Value"}, "total_sell_value": {"type": "number", "title": "Total Sell Value"}, "trade_count": {"type": "integer", "title": "Trade Count"}, "matched_symbols": {"type": "integer", "title": "Matched Symbols"}, "open_symbols": {"type": "integer", "title": "Open Symbols"}, "per_symbol": {"items": {"$ref": "#/components/schemas/SymbolPnLOut"}, "type": "array", "title": "Per Symbol"}}, "type": "object", "required": ["total_realized", "total_buy_value", "total_sell_value", "trade_count", "matched_symbols", "open_symbols", "per_symbol"], "title": "PnLOut"}, "PositionOut": {"properties": {"symbol": {"type": "string", "title": "Symbol"}, "side": {"type": "string", "title": "Side"}, "quantity": {"type": "integer", "title": "Quantity"}, "average_price": {"type": "number", "title": "Average Price"}, "last_price": {"type": "number", "title": "Last Price"}, "unrealized_pnl": {"type": "number", "title": "Unrealized Pnl"}}, "type": "object", "required": ["symbol", "side", "quantity", "average_price", "last_price", "unrealized_pnl"], "title": "PositionOut"}, "StateOut": {"properties": {"mode": {"type": "string", "title": "Mode"}, "live_armed": {"type": "boolean", "title": "Live Armed"}, "algo_state": {"type": "string", "title": "Algo State"}}, "type": "object", "required": ["mode", "live_armed", "algo_state"], "title": "StateOut"}, "SymbolPnLOut": {"properties": {"symbol": {"type": "string", "title": "Symbol"}, "buy_qty": {"type": "integer", "title": "Buy Qty"}, "sell_qty": {"type": "integer", "title": "Sell Qty"}, "avg_buy": {"type": "number", "title": "Avg Buy"}, "avg_sell": {"type": "number", "title": "Avg Sell"}, "net_qty": {"type": "integer", "title": "Net Qty"}, "realized_pnl": {"type": "number", "title": "Realized Pnl"}}, "type": "object", "required": ["symbol", "buy_qty", "sell_qty", "avg_buy", "avg_sell", "net_qty", "realized_pnl"], "title": "SymbolPnLOut"}, "TokenOut": {"properties": {"token": {"type": "string", "title": "Token"}}, "type": "object", "required": ["token"], "title": "TokenOut"}, "TradeOut": {"properties": {"time": {"type": "string", "title": "Time"}, "symbol": {"type": "string", "title": "Symbol"}, "side": {"type": "string", "title": "Side"}, "quantity": {"type": "integer", "title": "Quantity"}, "price": {"type": "number", "title": "Price"}}, "type": "object", "required": ["time", "symbol", "side", "quantity", "price"], "title": "TradeOut"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}, "input": {"title": "Input"}, "ctx": {"type": "object", "title": "Context"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}, "securitySchemes": {"HTTPBearer": {"type": "http", "scheme": "bearer"}}}}